Move index altering scripts to common lib
