+++ webapps/eea_search/scripts/removeData.js
@@ -1,4 +1,4 @@
-var api = require('./util/esAPI.js');
+var api = require('../../common/lib/esAPI.js');
 
 // As part of the runScript skel
 var run = function(options, callbacks) {
